pyproject.toml
[tool.poetry]
name = "personal-briefing"
version = "0.1.0"
description = "Daily AI-powered news briefing delivered to your inbox"
authors = ["Kenny Mroz Jr"]
readme = "README.md"
license = "Apache-2.0"
packages = [{include = "personal_briefing", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.0"
pydantic = "^2.6.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
moto = {extras = ["ses"], version = "^5.0.0"}
ruff = "^0.2.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
types-pyyaml = "^6.0.12"
types-requests = "^2.31.0"
boto3-stubs = {extras = ["bedrock-runtime", "ses"], version = "^1.34.0"}

[tool.poetry.group.cdk.dependencies]
aws-cdk-lib = "^2.130.0"
constructs = "^10.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=src/personal_briefing --cov-report=term-missing"
